Add timeout for NotificationListener CountDownLatch
PermissionGrantTest#grantLocationPermission_userNotified is flaky (about 7-8% of the time) and sometimes gives an error message containing "Timed out executing test <test_name> after 1798 seconds". Other flakes of this test offer a different explanation, though: "Test did not run due to instrumentation issue", which may be related to a CountDownLatch not having a timeout and thus forcing the environment to shut down the test (and possibly the whole run?).

This commit aims to make the reason of the flake consistent in order to identify it. And also to prevent the test from running for half an hour.

Bug: b/386323854
Flag: TEST_ONLY
Test: atest CtsDevicePolicyTestCases:android.devicepolicy.cts.PermissionGrantTest#grantLocationPermission_userNotified
Change-Id: I82a64afc7aaf78aee9a701bb8b3415d41235628d