Use first line of cmd overlay dump state output
When the current user is not the system user (userId 0), the output of
the cmd overlay dump state will include an additional line, specifing
the default overlays. This is due to the OverlayManagerService code
using the system user instead of the passed in user, where available.

This fixes the test utility by using the first line of the output only.

Flag: EXEMPT testfix
Bug: 391829821
Test: atest InputMethodServiceTest#testOnCustomImeSwitcherButtonRequestedVisible_gestureNav
  InputMethodServiceTest#testOnCustomImeSwitcherButtonRequestedVisible_threeButtonNav
Change-Id: Iaa1e807be47c8c28027fb0596d5f8cc8ee092ad1