Add additional verification for SetterCallback
Verify that the error code called from onErrorEvent is a known CPM error code. Verify that the error code has not been set previously and that the CoutDownLatch is still counting. Finally, assert that both a valid value and error code were not set via the callback.

Flag: EXEMPT test-only
Bug: 389826136
Test: atest CtsCarTestCases:CarPropertyManagerTest
Change-Id: I9075933d92b2822015b88df4d7172c918d0443c9