refactor test class
Changing the permissions of the test app while the app is bound to
will kill the test app. If the test app ever died, the rest of the
test would fail.

Now, permissions are changed on each test BEFORE binding to the
test app.  If there is no change in permissions or the process
is still alive, the rebind will be skipped.

Flag: EXEMPT, test only change
Fixes: 355720464
Test: refactor existing test class
Change-Id: I0825c878313b230e24004694e676bcfb539f5de7