Add delay after updating power property
When testing power dependent properties such as HVAC or ADAS
properties, when the power is turned on/off during the test, there
needs to be a delay to account for the vehicle updating the dependent
properties' status. By default this is set to 2000ms. This is notable
when verifying that get/set is not available after turning off the
HVAC.

Flag: EXEMPT test-only
Bug: 385810871
Test: atest CtsCarTestCases:CarPropertyManagerTest
Change-Id: I8397346518cbcd9680d5221e2d17ce4b6aa4fda5