Verify dependent property is not available for powered off area IDs
For ADAS and HVAC properties that are dependent on a power property such
as HVAC_POWER_ON, AUTOMATIC_EMERGENCY_BRAKING_ENABLED, etc, we turn
off/disable the feature and verify that the dependent property is not
available. However, it's possible turning off or disabling the feature
will fail for a valid reason (exception or error code) so we should only
verify for area IDs that successfully change the power state.

Flag: EXEMPT test-only
Bug: 385810871
Test: atest CtsCarTestCases:CarPropertyManagerTest
Change-Id: I54ece46173f649fa262041525f4128ad6fe27a5a