Update CTS regarding the HAL impl change
In ag/31417738, the default implementation for auto pq, auto aq and
super resolution is supported changed from false to true. So when
setting enable to the auto pq, auto aq and super resolution in the CTS,
it does change the enable to true. So when checking is those enabled,
the CTS will fail.

Bug: 391695285
Flag: EXEMPT bugfix
Test: atest VtsHalMediaQualityTargetTest
Test: atest CtsMediaQualityTestCases
Change-Id: I5fc4d76ef8288984acb5154df09c0b46f5bb4a1d