EvemuDevice: Use the first event's timestamp to mark end of descriptor
We send a no-op event to mark the end of the descriptor so that the
uinput command knows to register the device. Ensure that we use the
same timestamp as the first event in the file for the no-op event
to better support recordings taken with the --timestamp-base=epoch flag.

Bug: 385141667 
Bug: 389901828 
Change-Id: Iee2e60b7dd8a3e6fff8bde7cee16cb687d577c86
Test: Presubmit
Flag: TEST_ONLY