Make cts-install-lib test apps available to the platform
Soong was accidentally allowing a corner case where modules that
were not supposed to be available to the platform (due to
apex_available being set to an apex and not including
"//apex_available:platform") were being made available to the
platform as long as no apex modules depended on them.  Remove the
apex_available entires from test apps that are not in an apex so
they continue to get install rules.

Bug: 394664568
Test: compare out/soong/installs-${TARGET_PRODUCT}.mk
Change-Id: I9214d26eb8d55f464c2ab7812d333f632049f796