Remove existing role holder before adding another
PackageManagerShellCommandInstallTest that set and reset the current
test package as the role holder are currently flaking with process
crashed failures.

This could be because while resetting, if the previous role holder is
not null, we just add it right away without removing the test package
role holder. This is meant to replace the old role holder, but it could
also end up killing the old role holder - which is the test app in this
case, causing failures.

Bug: 390304109
Test: atest PackageManagerShellCommandInstallTest
Flag: TEST_ONLY
Change-Id: I9fac810edcb9c674f03b0467fe84340403f0cbb7