InputHidTestCase: Wait for additional device capabilities
Additional input device capabilities that are required by some tests,
like battery and vibrator, can be added to an InputDevice after
the device is initially added. Therefore, we use a polling check to
wait for the capability to be added, instead of expecting the
capability to be initially present.

Bug: 395870410
Change-Id: I36a163daf4fcef0838bfb641a178b751ae47d110
Test: Presubmit
Flag: TEST_ONLY