Add some simple media Output Switcher tests
These tests check that when an app is listening for routes with a
given feature and calls MediaRouter2#showSystemOutputSwitcher(), the
Output Switcher dialog shows routes advertised with that feature. It
also makes sure that clicking on one of those routes causes the route
provider service which advertised the route to get the appropriate
onCreateSession() callback.

Bug: 388361207
Test: atest OutputSwitcherTest
Flag: TEST_ONLY
Change-Id: Idabc5f1da95d1d18120f3e5f2ec385c4664fc4b9