Deflaky invokesCallback_cancelled
The onBackProgressed callback would be called after the threshold is
crossed and the TouchTracker become active, so there is no guarantee
that it would be called the same number of times as the motion event
occurs. Check progress should be more reliable.
Also extends the waiting time for assertInvoked, because the
onBackCancelled only be called until progress counts to 0.

Bug: 393570028
Test: atest OnBackInvokedCallbackGestureTest
Flag: EXEMPT bugfix
Change-Id: Iabf2b3e5256a9f2d9b24c4e4c7f4bd729c47dc65