Dsda: Resolve answering MT drops unholdable FG call
Add CUJ test coverage in order to verify the behavior of adding the
Connection#EXTRA_ANSWERING_DROPS_FG_CALL extra to the incoming call
connection extras when there's an ongoing unholdable call that would be
dropped from answering the call.

Also ensures that we don't include this extra for the managed unholdable
call and incoming self-managed call case so that we don't inadvertently
drop the carrier call in favor of the self-managed call.

Bug: 394900917
Test: atest CtsTelecomCujTestCases:CallSequencingBasicCallTest
Flag: com.android.server.telecom.flags.enable_call_sequencing
Change-Id: Ie87d2c1354d5f1e024061d09ca960c821037750d