Relax test_exposure_keys_consistent to account for missing metadata and post raw boost value.
OEMs are not required to offer the CONTROL_POST_RAW_SENSITIVITY_BOOST key, and
can also report a different value back in the CaptureResult if the HAL is unable
to apply the requested boost. Only verify that the two captures are roughly
equal if CONTROL_POST_RAW_SENSITIVITY_BOOST is being used incorrectly.

Bug: 396243502
Flag: EXEMPT ITS test
Test: Ran on physical device with exposure issue
Change-Id: I99b1aaa04e328ec6f7b9cf01bf8a3e7dad261a0e