Fixed the CTS test for local account.
CTS test expects the contacts are written to null account when account is set to null, and to the local account when account is not set. In fact, contacts should be written to the local account in both cases.

Bug: 397422668
Test: atest ContactsContract_DefaultAccountTest (running on a device
which manually overlay local account name and type config)
Flag: EXEMPTED

	modified:   contactsprovider/src/android/provider/cts/contacts/ContactsContract_DefaultAccountTest.java

Change-Id: I881d98323be777f738b8b88e04965f3592e8468e