Increase VirtualAudioPermissionTest resilience
On emulators/virtual devices it takes more to setup the VDM audio policy which is done in parallel with the startup of the app. Since apps are and should be unaware of the VDM audio redirection, there is no simple way to indicate to the app that the VDM virtual audio devices are "connected".

For now increase the test resilience until a more complex sync mechanism is devised. Value chosen based on multiple manual runs.

Bug: 397375912
Test: atest CtsVirtualDevicesAudioTestCases_audio_virtualaudiopermissiontest --iterations=50 (on aosp_cf_x86_64_only_phone-trunk_staging-userdebug)
Flag: TEST_ONLY

Change-Id: Ib5996ed9bc7236f4cb8cdb79e90dcb6ce16e146d