[Autofill CTS] Add in assertion first fill request is received to de-flake test.
testResponseFirstAfterResponseSecond() is currently flaky because after
the first fill request, it only verifies the InstumentedAutofillService
onConnect() is called before continuing to trigger next request. However
this doesn't guarantee the first onFillRequest() is called on
InstrumentAutofillService. It's possible the second request cancels the
first request before first onFillRequest() is called. This leads the
test sometimes to fail, because originally the first CannedResponse is
prepared for first fill request, now it will reply to the second fill
request.

This change takes one step further to assert the onFillRequest() is
called on InstrumentAutofillService before continuing on to trigger next
fill request. Calling this added method
waitUntilFirstFillRequestReceived() will wait until the first fill
request is received after onConnect() is called on
InstumentedAutofillService.

Without this change, current flaky rate on
cf_x86_64_phone-trunk_staging-userdebug is 3%:
https://android-build.corp.google.com/test_investigate/invocation/I65400010346851451/test/TR63929555105415645/?itab=tests

With this change, run on ABTD shows run passes 100/100 times on ABTD:
https://android-build.corp.google.com/test_investigate/invocation/I27600010362501504/test/TR56529615032895761/?redirect=http%3A%2F%2Ffusion2%2Fb55c9a13-92d9-41bc-8eb9-c79abede6105&status=none

Flag: TEST_ONLY
Bug: 394737411, 396513948
Change-Id: I41a90c1e6ab98c91b53dd0e6cd81f1fa23efb23f
Test: atest CtsAutoFillServiceTestCases