Don't fail on package preferences with UID 0
There have been cases in practice where some package preferences are associated with UID 0 and causing this test to fail. PreferencesHelper just reports the UID with the data that it gets, so even if that UID is in error somehow, the statsd atom tests shouldn't fail just for reporting what data it was given.

Fixes: 391459856
Flag: EXEMPT bug fix
Test: this
Change-Id: I034041a73356bfe7634eafd0c7a6528ca54856eb