RESTRICT AUTOMERGE: Use activity window screenshot not whole device.
ActivityTransitionTests uses screenshotting to validate visual changes
during activity transitions. This change limits the screenshot to the
test activity for more hermetic testing on multi-window freeform
environments where a full screenshot will contain everything on the
display not just what the test needs to validate.

This is a redo of I1f610da0911a8afb019825d505ae0152dabee9d0 differing from the previous change in that a "fallback" screenshot mechanism is incorporated that uses both UiAutomation screenshot as well as WindowManager#snapshotTaskForRecents for validation - see the fallbackScreenshot portion of this CL.

Test: android.server.wm.ActivityTransitionTests
Bug: 349156941
Change-Id: I75e7653ef45aa3c08282ba4c1c48d4e5791a7d51