[CTS] Ensure Notification Verifier Tests button visibility
Buttons are disabled on any config change since the setup is not invoked. This change makes sure the button visibility. We also moved RTL tests at the end of the flow and put a note that tells testers to disable the RTL mode at the end of the test case.

Bug: 395773597
Flag: android.app.api_rich_ongoing
Test: open Notification CTS-Verifier and change device config (e.g. Rotate the screen, change RTL settings, etc.)
Change-Id: I0515deb119355dd37f57f18b7c320fcf0c5d9811