Run device owner broadcast test on system user
The broadcast that device owner changed is only sent to the user that is
set as device owner. A test that sets a DO on the system user hence
needs to run on the system user.
This worked fine for non-HSUM because by default tests are only run on
the system user, but for HSUM the default is user 10, by adding the
annotation to run on system user, the test also only runs on user 0 on
HSUM.

It seems a bit strange that bedstead would by default set a DO on user 0
on HSUM, and test with affiliated mode. Affiliated mode isn't used. The
DO mode that is used at least for AL teamfood at the moment is DO on
user 10. It'd probably be better to default to DO on user 10, but would
cause a bunch of tests to fail.

Bug: 397070443
Test: atest android.devicepolicy.cts.DeviceOwnerTest#setDeviceOwner_receivesOwnerChangedBroadcast
Flag: EXEMPT TEST_ONLY
Change-Id: Ia54762839a3dfab27a0fc98738ca1e02f68b04ae