Write dumpOnFailure to app files
The DumpOnFailure utility was trying to write to /sdcard/DumpOnFailure,
which would fail if the test app did not have the permission for
external storage. Instead, we can simply write to the files of the test
app, and pull from there, removing and requirements on permissions.

This also removes the manual cleanup of these files, instead delegating
it to the FilePullerLogCollector, which will only clean up after
successfully reading the files.

Flag: EXEMPT testfix
Bug: 395783236
Test: atest ActivityTransitionTests
  BlurTests
  SplashscreenTests
  EnsureBarContrastTest
Change-Id: Ic024d378223d1278808d179df0ed3a26026f5f24