Avoid assuming built-in speaker is media output
OTT devices (sticker, stb) often do not contain build-in speaker and are expected to always connect to HDMI. So, instead of assuming
the default output is the built-in speaker, require that a route
with the default id is not present. As that would be indicative of
lack of permissions.

Bug: 398153597
Test: build and test on stb device
Flag: EXEMPT test fix
Change-Id: I78ae2e34f90313b6899556bd07f730a01153ecc6