Write to external storage for DumpOnFailure
This is a partial revert of [1], to write the DumpOnFailure files again
in external storage. Reading them with the FilePullerLogCollector from
the app files only works if the device is rooted, which will not be the
case on CTS tests.

This will instead write them to the directory given by
getExternalStorageDirectory. While the test config has the path of
"/sdcard/..." for reading the files, this will get resolved to the
actual external storage path.

  [1]: Ic024d378223d1278808d179df0ed3a26026f5f24

Flag: EXEMPT testfix
Bug: 399050749
Test: atest ActivityTransitionTests
  BlurTests
  SplashscreenTests
  EnsureBarContrastTest
Change-Id: I83dcce8a0e71d6cad01e749561e2980aaa4d63dd