Fix flaky CTS DynamicInstrumentationManagerTest
The location of `services.odex` is relatively stable on device, but ART sometimes moves this location to a cache path with a slightly different naming convention.

Update the test to be less flaky, as the exact path and filename is not consequential to what the subject test is about.

Bug: 398997714
Change-Id: I7aeb97e84c731914e1f8d67b5751514d92b9f724
Test: TH
Flag: EXEMPT test fix