Catch exception from isPropertyAvailable.
isPropertyAvailable is not supposed to throw exception, however, due
to an implementation issue in Android 14 and 15, it may throw
IllegalArgumentException if the property is not supported. In the
CTS test we should consider this case and skip the test accordingly.

Test: atest android.car.cts.CarPropertyManagerTest
Flag: EXEMPT test fix
Bug: 398768258
Merged-In: I8ebfec5523cadd620aeaa88529b0492fec59374b
Change-Id: I8ebfec5523cadd620aeaa88529b0492fec59374b