Use the status checker app in testSdmOk.
The status checker app enables checking the compiler filter /
compilation reason on itself, where the actual values are returned by
the runtime, to check whether the runtime is really using the
compilation artifacts.

The actual check will be added later, when the platform support for
Cloud Compilation has been fully implemented.

Installing the status checker app during the test means we cannot
install it in the setup phase through SuiteApkInstaller, so this CL
removes SuiteApkInstaller.

Bug: 377474232
Test: atest CtsCompilationTestCases
Change-Id: Ic4157b03a0b288ed86a9a6cf33286d0989758b14