Do not launch home activity after entering PIN
For some reason, launching the home activity interferes with dismissing
the Keyguard on Automotive, causing the device to never actually become
unlocked after entering the PIN.

The purpose of launching the home activity immediately after entering
the PIN probably was to move any other activity to the background.
However, that is unnecessary for just testing device lock and unlock.
It might be a bug that launching the home activity interferes with
dismissing the Keyguard, but it's unclear how it could be user
impacting, and in any case it isn't what the Keystore tests are intended
to test; they just do what is needed to lock and unlock the device.

Therefore, stop launching the home activity in DeviceLockSession.

Bug: 393171575
Test: atest CtsKeystoreTestCases
Test: atest --rerun-until-failure=10 CtsKeystoreTestCases:android.keystore.cts.CipherTest#testKeyguardLockAndUnlock
Test: atest --rerun-until-failure=10 CtsKeystoreTestCases:android.keystore.cts.CipherTest#testEmptyPlaintextEncryptsAndDecryptsWhenUnlockedRequired
Flag: TEST_ONLY
Change-Id: I3d1e017d66b84e1180428358f20c73918d1cad69