Clear the helper apps from "bad" processes list.
Some of the tests are verifying that the helper app
cannot start FGS from BG and that results in the app
crashing. If the tests are ordered in such a way that
the helper app end up crashing multiple times in quick
succession, the app could end up on the "bad" processes
list. This will result in app not being able to start
in some cases. So, explicitly clear the helper apps
from bad processes list at the start of each test to
avoid this scenario.

Bug: 401276473
Test: atest tests/app/src/android/app/cts/ActivityManagerTest.java
Flag: TEST_ONLY
Change-Id: I3fcd340aefff3d8c529f22b177a293c2c83383c5