Fix vendor API level gate for attested Verified Boot key check.
The vendor API level format changed in Android 14-QPR3 since the
SDK API level and vendor API level became decoupled. This check
was introduced in aosp/3344782 with the intention of running on
VSR-16+ devices, but will actually end up running on VSR-15+
devices (since the vendor API level for Android 15 is 202404,
which is >= 36).

Testing note: all tests that execute as part of the "atest" command
below pass except two ("test{Ec,Rsa}Attestation_DeviceLocked"), which
is expected since the tests run on an unlocked device.

Bug: 384021501
Bug: 398693391
Test: atest CtsKeystoreTestCases:android.keystore.cts.KeyAttestationTest
Change-Id: I47578685d99afaffc68f650261638e3c87393af9