Remove AHardwareBufferTest_WithoutCPUAccessLock
This test relies on the vendor gralloc implementation to reject the lock
request if the underlying AHB was allocated with no cpu access bits.
However, not every gralloc implementation actually does this, so VF
devices end up failing this test.

Will reintroduce in a different way in a future release.

Bug: b/385240723
Bug: b/399400164
Flag: TEST_ONLY
Change-Id: I948b1cbdd17477fc12e52df9eb053697b911d4aa