Wait for the popupwindow to be on top before clicking
The test used to click on the popup window without ensuring that it was
on top. This caused test flakiness. Locally, the test used to fail when
executed as part of the larger test.

Instead, in this CL we update the logic to ensure that the popup window
is fully on top before injecting.

Bug: 402475691
Test: atest android.view.cts.TooltipTest
Test: atest CtsViewTestCases_input_related
Flag: TEST_ONLY
Change-Id: Ia43fb626d8ba33e336891cb636ebf6e1f4480faf