Fix flaky wear preference test
When the test scrolls to look for certain views, the view may be skipped
if the system is running slow or laggy. This fix changes the scrolling
logic to more incrementally scroll so there is a lower chance to skip
the view.

Bug: 397728119
Test: atest
android.preference.cts.PreferenceActivityFlowPortraitTest#listDialogPortraitTest
on cf_gwear_x86-trunk_staging-userdebug
Flag: EXEMPT test_only

Change-Id: I85d0b8f03fc93e20d5314543fbc3dfe2825005fe