Check TestActivity focus after back press for IME
If the IME back animation controller is not registered as a callback,
the back press won't be processed by the IME, and will instead be sent
to the TestActivity, which will be dismissed. This won't immediately
fail the test, as the IME does get hidden through the activity being
dismissed.

This adds an explicit assertion that the TestActivity is still focused
(and thus not dismissed) after the back press, to provide a stronger
signal for the failure.

Flag: EXEMPT testfix
Test: atest KeyboardVisibilityControlTest#testIMEVisibleInSplitScreenWithWindowInsetsApi
Bug: 402671709
Change-Id: I4ae89db6d5ecae1e22dd1cd5f4767b00aa0231d5