[audio] Partial fix AudioManagerTest vol init
When initializing the volume for all streams, if one of the streams is
aliased, we can change the volume of its alias away from INIT_VOL, if
the two streams have sufficiently different ranges.

Avoid this for now by skipping setting the volume for all aliased
streams. The aliased streams volumes will be consistent between runs but
depend on the range of the alias. Most tests don't rely on the INIT_VOL
for a potentially aliased stream, but this should be fixed more durably.

Skip testAccessibilityVolume if it is aliased, for this reason.

Test: atest AudioManagerTest
Flag: EXEMPT test fix
Bug: 397350485
Change-Id: I019d048a513c1c7f834b1d6797d1d229110e1c0c