cleanup: Update A11yE2ETest to retry all gesture dispatch.
If the gesture dispatch is triggered immediately after an A11yService
updates its A11yServiceInfo, then the dispatch may get cancelled
automatically if the A11yInputFilter is rebuilt in the middle of
dispatch. This can be resolved by a 1-time retry.

We don't currently offer a way for a CTS test to intelligently wait
for A11yInputFilter to finish rebuilding itself, as this is an internal
implementation detail, so this one-time retry (which happens ~7% of
the time) is more efficient than adding a sleep that would delay the
test 100% of the time.

Fix: 403694040
Test: use abtd to run the failing tests 100 times
Flag: EXEMPT test cleanup
Change-Id: Ied394c7b848303ed9d0f47cd850c0688928985dd