Add more logs for input devices listener
I'm observing a presubmit failure where the device added notification is
not received within 20 seconds, even though the logs show that the
InputManagerGlobal received "device added" call.

That means that either listener wasn't registered, or one of the checks
inside the listener failed. Add a log to help figure out which one it
is.

Bug: 406273885
Test: atest InputTests
Flag: TEST_ONLY
Change-Id: I9158088d4e0e7ad8748eb18a303bf1cc5b83d7bd