Fix flaky emergency call CUJ test
Some CUJ tests are flaky due to a recent fix to prevent waiting on the
live call to disconnect (for single sim scenarios) before placing the
emergency call. With the way that the test ManagedConnectionService is
set up, it will fail the new outgoing connection if there is more than
1 ongoing call. Depending on when the live call does get disconnected,
we can see flakiness in some test runs. Instead, ensure that we
explicitly disconnect any active calls and don't fail the emergency
call.

Bug: 405920289
Flag: EXEMPT test fix
Test: atest CallSequencingMultiCallEmergencyCallTest
Change-Id: I65889868c286b0a2c0485b16cc5f3c7840ff9dc9