Use system property for screenshot size on XR
This CL's cherry-picked from aosp/3566142 for migrating A14 CTS to A16, 1st step is to migrate to internal main.

This tests assumes that the screenshot resolution will be equal to the
display resolution. This assumption is false on AndroidXR. This change
introduces a modification to how the expected screenshot resolution is
determined. If the platform is XR, then a new system property will be
queried. If it exists, then it will parse and use that property as the
expected values for the screenshot resolution. If the property does not
exist, it will fallback to using Display.getRealSize. All other
platforms will continue to use Display.getRealSize.

Bug: 361007807
Test: Verify test passes on AndroidXR
Test: Verify tests pass on non-XR platforms
Change-Id: I28ba3d9d0e1004062b284131d6c496951d472199