Removing tests for exclusive use of /dev/usb_accessory
The failures seem to be the result of a kernel change that recently
removed exclusive access to the accessory device. (for more info, refer
b/399308334). The test passed successfully earlier because it was
consistent with older kernel behaviour. In absence of other approaches,
this CL removes the few test lines that test for this behaviour which
cause this failure and push that change to prevent further issues. More
details: old kernel = <6.6, new kernel = <=6.6

TEST: manual
BUG: 399308334
BUG: 396560270
FLAG: TEST_ONLY

Change-Id: I1f874cd7fccc391a8caff0d2fb50a944ed453a8d