Use a11y settings to extend toast timeout in tests
This will allow the toast to show on screen longer, thus helping us
avoid test flakiness.

Our team is finding that even if 3.5 second duration is requested for
the toast, sometimes the toast can be on screen for less than one
second. This is causing occasional test flakiness.

Bug: 405221087
Bug: 407551227
Test: atest WindowUntrustedTouchTest#testWhenOneCustomToastWindowAndOneSawWindowBelowThresholdFromDifferentApp_blocksTouch --iterations 10
Flag: TEST_ONLY
Change-Id: I7d222c0b8d9ce00ea08f517072fe0a2fd68df599