Fix issue with RingerTest reliability.
The testIncomingCallVibrateMode_vibrationSettingsSupported_VibrateAndRing
test was not running reliably when I was testing b/308210808.
The root cause was that the audio playback callback was coming back with
empty results.  It needs to be registered and unregistered as shell
identity to ensure the data is populated in the callback.

FLAG: NONE test only fix
Test: This is a test
Bug: 308210808
Change-Id: I227031b3cb23c56f2cce7272abc9dbe93b1c580a