Wait for PiP config callbacks to enter
When we enter PiP, we send a config change to the client
at the end of the transition, so the client would receive
the actual onPictureInPictureModeChanged() callback after
the transition is finished (and after the WM state is updated).

So if we are interested in counting the number config changes
for an interaction after, like removing PiP here, we should wait for
that first config-change client call before doing separateTestJournal().
This helps avoid the race condition that would count 2 PiP mode change
callbacks instead of a single one.

It looks like this would have been flaky in PiP1 too before; it's just
that the test was off when PiP2 went into staging at first, which is why
ATI would show 0% flake rate for a while.

Bug: 405360172
Flag: EXEMPT bugfix
Test: atest PinnedStackTests#testStopBeforeMultiWindowCallbacksOnDismiss
Change-Id: I4c3b8bd2aa6c824440cca4138b701e005b8dde6b