Ignore Cronet's TraceEvent class in ApiTests
This class requires prior initialization which this test does not know
how to perform. Normally, this is fine; as this test tries to recover
from these scenarios by catching all exceptions. Unfortunately, for
this specific case, this results in the static constructor of this class
JNI'ing into native code with no well defined state. This leads to a
native crash, which the test code cannot recover from.

Test: atest CtsHiddenApiKillswitchWildcardTestCases
Flag: TEST_ONLY
Bug: 407596762
Change-Id: I6c7d0a8ccd118487086bff137f679a2dfda9d2bc