Fix failed BackgroundActivityLaunchTest for ScallableUI
In some tests, it starts the testing activity, waits for it focused,
then sends a Broadcast to the testing activity. On ScallableUI build
when the activitiy is focused, its onCreate() callback is not invoked
yet, thus its BroadcastReceiver is not registered yet, so the testing
activity never receives the Broadcast.

To fix it, this CL changed to it wait for the testing activity resumed.

Fixes: 402786130
Test: on ScallableUI build, run: atest android.server.wm.BackgroundActivityLaunchTest --iterations 5
Flag: EXEMPT bug fix
Change-Id: Idff91c4a7839443eb1e7cba058b57b4d0f22e09f