DO NOT MERGE Fix missing Hover events.
Injected Hover mouse events is sometimes lost and
not delivered to few (regardless of wait). To remedy
this the Hoover event is send again (max 5 times) if
the target list item does not get selected. There is
also a flakyness in the scrolling to end of list that
also causes this test to fail.

Test: atest CtsWidgetTestCases:android.widget.cts.PopupMenuTest#
testHoverSelectsMenuItem

Bug: 408237223
Change-Id: I51703d8838f967a988c00f9a3c27adf680334049