Fix android.virtualdevice.cts.camera.VirtualCameraCaptureTest.inputRate_HigherThanMaxFps_allFulfilled
We could not rely on the sensor timestamp to measure the time it takes
to capture frames since it is artificially written by the image writer.

Instead we now measure the time between two calls to onCaptureCompleted
which is a good approximation of the time the capture took.

We use the average since verifying every single frame does not produce
stable results for the test.

Test: android.virtualdevice.cts.camera.VirtualCameraCaptureTest.inputRate_HigherThanMaxFps_allFulfilled
Flag: TEST_ONLY
Bug: 406965817
Change-Id: Ic9ddf994d8661f4dd9c1334514c346879153f26a