Reset animation scale overrides for AttachedSurfaceControlSyncTest and SurfaceViewSyncTest tests
Based on failed test logs, it looks like the test activities are not
generating any frames causing the tests to fail. The tests are dependent
on animation scales, so add class rules to ensure any animation scale
overrides are reset before the test is run as a speculative fix.

Test: presubmit
Bug: 406693295
Flag: EXEMPT test fix
Change-Id: I39347bf1b3590f56ac2eae56bdeb0b75f3d1a733