mediav2: parse performance point string for width and height
performance point class does not have api that allows reading width
and height information. parse string for the same.

MediaCodec.onRequiredResourcesChanged() callback is triggered only when
the resources required to process the media resource/resources is
updated after the initial estimation. This need not be triggered at the
start of the session after MediaCodec.configure(). The test is updated
accordingly.

Bug: 384051709
Test: atest CtsMediaV2TestCases:VideoDecoderAvailabilityTest
Test: atest CtsMediaV2TestCases:VideoEncoderAvailabilityTest
Flag: TEST_ONLY

Change-Id: Icf7efaae302242346ab45000a0d2e0263baae8da