RESTRICT AUTOMERGE: Add test condition in CTS-Verifier manifest for excluding Tap2Tone test in XR environment.
The "Audio Tap To Tone" test determines the time required for audio triggered by a user action on the screen to appear at the output of the audio system. In XR environment, user can not produce a "Tap" sound to trigger tap to phone test. The testing condition will never be satisfied. So this test case should be disabled in XR devices.

Bug: 390190799
Test: local test with emulator
Change-Id: Ia858b02ab1073383054a6c45ee2622070a87d1af