[cts] don't reuse the same broadcast receiver for uninstall
The same broadcast receiver is used to receive installation result, and it immediately unregisters itself after receiving the result. During the uninstall test, the same broadcast receiver can be interfered by other ongoing installation events, causing the uninstall results to be lost and the test to timeout. This CL makes a separate broadcast receiver for uninstalls to avoid such interference.

BUG: 363659880
FIXES: 363659880
Test: presubmit
FLAG: TEST_ONLY
Change-Id: Ife5483776dc3e910d74d956c8880fdfda6cc64b0