mediav2: use right media format during decoder selection
some tests of ctsmediav2testcases suite are validating the results of
encoder by decoding them with available decoders on the device. To
determine the available decoders, the media format used is same as the
one used for encoder configuration. While it seems acceptable, there is
a major flaw. Not all keys are required to be honored by encoder. For
instance profile key is treated as a suggestion by the component and can
be updated to a compatible and supported profile. The level information
configured can be incorrect and can be updated basing on the encoding
parameters. Otherwords output format may not be always exactly identical
to input format but could be a compatible variant of it. Hence, using
input format to obtain supported decoders list can result in no decoders
available thus failing the test. This change addresses this.

Bug: 409619775
Bug: 409951220
Test: atest android.mediav2.cts.EncoderHDRInfoTest
Test: atest android.mediav2.cts.EncoderColorAspectsTest
Flag: TEST_ONLY

Change-Id: I1ebebc321c94960fd769aef47760f9a40a3edc41