Catch security exception with adjustStreamVolume
If a security exception is thrown from invoking
AudioManager#adjustStreamVolume because the device is in DND, then we
will not be able to run the tests. This was causing an uncaught security
exception to be thrown which puts the subsequent tests into a bad state
because tearDown is not properly called.

Bug: 409634974
Flag: EXEMPT test fix
Test: atest BackgroundCallAudioTest
Change-Id: I4284b382d24ac94e7957d0888ac6eaef77a07031