Add fallback support for checking tablet in CTS.
Not all devices set ro.build.characteristics, so we have to use more creative ways to figure out if it's a tablet.

Check if the device is a foldable, and if so, it's not a tablet.

Otherwise check if the screen size is too large to not be a tablet.

Docs on figuring out of the device is a phone/foldable/tablet: https://android-developers.googleblog.com/2023/06/detecting-if-device-is-foldable-tablet.html.

Bug: 406234957
Test: EXEMPT feature flag string change
Flag: EXEMPT test util only change
Change-Id: I921a8ad970540a950991343c64549d67267202c1