SELinuxHostTest: More effectively use caches
getDeviceFile() unconditionally called to the remote device
to verify the existance of a file, prior to checking the cache.
This defeats the purpose of the cache and drastically slows down
test performance. Move the existance check after the cache check.

Add a new cache for isSepolicySplit(). This avoid a round-trip
request to the device per test.

These two changes result in a significant speedup in execution
time.

  Command:
    time atest CtsSecurityHostTestCases:android.security.cts.SELinuxNeverallowRulesTest
  Before this patch: 1014 seconds
  After this patch: 487 seconds

Test: atest CtsSecurityHostTestCases
Flag: TEST_ONLY
Bug: 407308692
Bug: 407295122
Change-Id: I02e4f19ffe3dd5269f9687ebc9b9a27b0245ea7e